Modify c4_setTempDir and eliminate FilePath::tempDirectory
Motivation:  There is no place to have a default temporary directory that is guaranteed to be on the same filesystem as it needs to be (move operations require the same filesystem to function)

Main changes:  c4Base.h c4_setTempDir signature has changed, and it is now an error to call it twice.

The rest of the commit happened because the tests were still using the tempDirectory call that I removed, and moving that logic up to the test level plowed me into some technical debt regarding `using namespace std` in headers which causes frustrating compile errors (especially on Windows).  There are a few fixups, including a Fleece bump because of an incompatibility I noticed.  The vast majority is just moving the using statement out of the headers and into the compilation files instead, and qualifying things as needed.